Mythril 0.1
Welcome everyone! This is the first public version of Mythril, A cozy, playful, and seamless way to explore the web. With an intuitive design, smooth transitions, and subtle creativity, Mythril makes every click feel delightful. In this version the basics have been added. We hope to inspire you all, as this is not a fully usable version. It has almost no protection from threats, and is purely built on webkit with simple features. Mythril does not collect any data, as it doesn't even have a server to call back to. 

Discover the web, your way.